+++ Details/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Details! Damage Meter
 
-## [Details.20260619.15236.172](https://github.com/Tercioo/Details-Damage-Meter/tree/Details.20260619.15236.172) (2026-06-19)
-[Full Changelog](https://github.com/Tercioo/Details-Damage-Meter/compare/Details.20260618.15235.172...Details.20260619.15236.172) 
+## [Details.20260620.15237.172](https://github.com/Tercioo/Details-Damage-Meter/tree/Details.20260620.15237.172) (2026-06-20)
+[Full Changelog](https://github.com/Tercioo/Details-Damage-Meter/compare/Details.20260619.15236.172...Details.20260620.15237.172) 
 
-- Don't use toc cache values, check for game version each iteration  
+-  Quick test